move planning modules obligations, plan, and probes into codestory-agent
